fix deposit-tets.c file2

+++ test/deposit-test.c
@@ -11,106 +11,106 @@ CTEST(input_check, test_input)
 
 CTEST(calc_test, calc)
 {
-    double result = Dohod(123400,140);
+    int result = Dohod(123400,140);
     
-    const double expected = 133272;
+    const int expected = 133272;
     ASSERT_EQUAL(expected, result);
 }
 
 CTEST(calc_less0_days, calc)
 {
-    double result = Dohod(22222, -5);
+    int result = Dohod(22222, -5);
 
-    const double expected = 0;
+    const int expected = 0;
 
     ASSERT_EQUAL(expected, result);
 }
 
 CTEST(calc_more365_days, calc)
 {
-    double result = Dohod(22222, 366);
+    int result = Dohod(22222, 366);
 
-    const double expected = 0;
+    const int expected = 0;
 
     ASSERT_EQUAL(expected, result);
 }
 /////////////////////////////////
 CTEST(calc_31_day_less_100000_rub, calc)
 {
-    double result = Dohod(50000, 31);
-    const double exp = 51000;
+    int result = Dohod(50000, 31);
+    const int exp = 51000;
 
     ASSERT_EQUAL(exp, result);
 }
 
 CTEST(calc_31_day_more_100000_rub, calc)
 {
-    double result = Dohod(150000, 31);
-    const double exp = 154500;
+    int result = Dohod(150000, 31);
+    const int exp = 154500;
 
     ASSERT_EQUAL(exp, result);
 }
 /////////////////////////////////
 CTEST(calc_120_day_less_100000_rub, calc)
 {
-    double result = Dohod(50000, 31);
-    const double exp = 51000;
+    int result = Dohod(50000, 31);
+    const int exp = 51000;
 
     ASSERT_EQUAL(exp, result);
 }
 
 CTEST(calc_120_day_more_100000_rub, calc)
 {
-    double result = Dohod(150000, 31);
-    const double exp = 154500;
+    int result = Dohod(150000, 31);
+    const int exp = 154500;
 
     ASSERT_EQUAL(exp, result);
 }
 /////////////////////////////////
 CTEST(calc_121_day_less_100000_rub, calc)
 {
-    double result = Dohod(50000, 121);
-    const double exp = 53000;
+    int result = Dohod(50000, 121);
+    const int exp = 53000;
 
     ASSERT_EQUAL(exp, result);
 }
 
 CTEST(calc_121_day_more_100000_rub, calc)
 {
-    double result = Dohod(150000, 121);
-    const double exp = 162000;
+    int result = Dohod(150000, 121);
+    const int exp = 162000;
 
     ASSERT_EQUAL(exp, result);
 }
 /////////////////////////////////
 CTEST(calc_240_day_less_100000_rub, calc)
 {
-    double result = Dohod(50000, 121);
-    const double exp = 53000;
+    int result = Dohod(50000, 121);
+    const int exp = 53000;
 
     ASSERT_EQUAL(exp, result);
 }
 
 CTEST(calc_240_day_more_100000_rub, calc)
 {
-    double result = Dohod(150000, 121);
-    const double exp = 162000;
+    int result = Dohod(150000, 121);
+    const int exp = 162000;
 
     ASSERT_EQUAL(exp, result);
 }
 /////////////////////////////////
 CTEST(calc_241_day_less_100000_rub, calc)
 {
-    double result = Dohod(50000, 241);
-    const double exp = 56000;
+    int result = Dohod(50000, 241);
+    const int exp = 56000;
 
     ASSERT_EQUAL(exp, result);
 }
 
 CTEST(calc_241_day_more_100000_rub, calc)
 {
-    double result = Dohod(150000, 241);
-    const double exp = 172500;
+    int result = Dohod(150000, 241);
+    const int exp = 172500;
 
     ASSERT_EQUAL(exp, result);
 }
@@ -118,16 +118,16 @@ CTEST(calc_241_day_more_100000_rub, calc)
 
 CTEST(calc_365_day_less_100000_rub, calc)
 {
-    double result = Dohod(50000, 241);
-    const double exp = 56000;
+    int result = Dohod(50000, 241);
+    const int exp = 56000;
 
     ASSERT_EQUAL(exp, result);
 }
 
 CTEST(calc_365_day_more_100000_rub, calc)
 {
-    double result = Dohod(150000, 241);
-    const double exp = 172500;
+    int result = Dohod(150000, 241);
+    const int exp = 172500;
 
     ASSERT_EQUAL(exp, result);
 }
